Refactors utilities for triangulation
Moves `find_extreme_coordinates` from `predicates.rs` to a new `utilities.rs` file in the `delaunay_core` module.
This change improves code organization and promotes reusability of utility functions within the core triangulation logic.
